Buildable with Visual C++ 2017

+++ asip.cpp
@@ -227,7 +227,7 @@ Side ASIP::sideToMove() const
   return toSide(get<QString>("turn")[0].toLatin1());
 }
 
-std::pair<GameTreeNode,unsigned int> ASIP::getMoves() const
+std::pair<GameTreeNode,size_t> ASIP::getMoves() const
 {
   const QReadLocker readLocker(&mostRecentData_mutex);
   return toTree(get<QString>("moves").toStdString());
